requirements.txt
torch
torchvision
transformers
Pillow
tqdm
python-Levenshtein
